+++ lib/src/platform_detector.dart
@@ -25,5 +25,5 @@ class PlatformDetector {
   }
 
   /// Check if the current platform is supported by this package
-  static bool get isSupported => usesFlutterWebView || isWindows;
+  static bool get isSupported => usesFlutterWebView || isWindows || isWeb;
 }