+++ lib/src/webview_factory.dart
@@ -1,2 +1,5 @@
 export 'webview_common.dart';
-export 'webview_impl_io.dart';
+export 'webview_impl_default.dart'
+    if (dart.library.js_interop) 'webview_impl_web.dart'
+    if (dart.library.html) 'webview_impl_web.dart'
+    if (dart.library.io) 'webview_impl_io.dart';