+++ lib/src/webview_impl_io.dart
@@ -1,17 +1,15 @@
-
 import 'dart:io';
-import 'package:flutter/foundation.dart';
 import 'webview_common.dart';
-import 'webview_impl_mobile.dart' as mobile;
-import 'webview_impl_windows.dart' as windows;
+import 'webview_impl_mobile.dart';
+import 'webview_impl_windows.dart' as win;
 
-EChartsWebView createWebView() {
-  if (kIsWeb) {
-    throw UnimplementedError('Web support is not available for this package.');
-  } else if (Platform.isWindows) {
-    return windows.EChartsWebViewWindows();
-  } else if (Platform.isAndroid || Platform.isIOS || Platform.isMacOS) {
-    return mobile.EChartsWebViewMobile();
+Future<EChartsWebView> createWebView() async {
+  if (Platform.isWindows) {
+    return win.EChartsWebViewWindows();
   }
-  throw UnimplementedError('Platform not supported: ${Platform.operatingSystem}');
+  if (Platform.isAndroid || Platform.isIOS || Platform.isMacOS) {
+    return EChartsWebViewMobile();
+  }
+  throw UnimplementedError(
+      'Platform not supported: ${Platform.operatingSystem}');
 }